requirements.txt
streamlit
google-cloud-bigquery
openai
openai-agents